+++ src/aks-preview/azext_aks_preview/commands.py
@@ -148,10 +148,6 @@ def load_command_table(self, _):
         g.custom_command('get-os-options', 'aks_get_os_options')
         g.custom_command('operation-abort', 'aks_operation_abort', supports_no_wait=True)
 
-    # AKS Copilot commands
-    with self.command_group('aks') as g:
-        g.custom_command('copilot', 'start_chat')
-
     # AKS maintenance configuration commands
     with self.command_group('aks maintenanceconfiguration', maintenance_configuration_sdk, client_factory=cf_maintenance_configurations) as g:
         g.custom_command('list', 'aks_maintenanceconfiguration_list')
@@ -269,8 +265,17 @@ def load_command_table(self, _):
             'aks_mesh_enable_ingress_gateway',
             supports_no_wait=True)
         g.custom_command(
+            'enable-egress-gateway',
+            'aks_mesh_enable_egress_gateway',
+            supports_no_wait=True)
+        g.custom_command(
             'disable-ingress-gateway',
             'aks_mesh_disable_ingress_gateway',
+            supports_no_wait=True,
+            confirmation=True)
+        g.custom_command(
+            'disable-egress-gateway',
+            'aks_mesh_disable_egress_gateway',
             supports_no_wait=True,
             confirmation=True)
         g.custom_command(